ctr reuse Client's GetGRPCConnection
